Move 클래스 생성
Player Move 클래스로 플레이어 앞/뒤, 왼쪽/오른쪽 이동과 드리프트 로직까지 이동시켜 분할, 나머지 로직들도 분할 예정, 좀 더 다듬어야 함(미니맵은 버그 다 고치고)

+++ PortFolio/Assets/04 Scripts/Controller/Game/WheelController.cs
@@ -9,11 +9,11 @@ public class WheelController : MonoBehaviour
 
     public WheelFrictionCurve ForwardFriction { get { return m_wheelCollider.forwardFriction; } }
     public WheelFrictionCurve SideWayFriction { get { return m_wheelCollider.sidewaysFriction; } }
-    public void Move(float speed, float dir)
+    public void Move(float speed, int dir)
     {
         m_wheelCollider.motorTorque = speed * dir * Time.fixedDeltaTime;
     }
-    public void Turn(float speed, float dir)
+    public void Turn(float speed, int dir)
     {
         m_wheelCollider.steerAngle = speed * dir * 10f * Time.fixedDeltaTime;
     }
